shellcheck

+++ common.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+bold=$(tput bold)
+blue="\e[0;94m"
+yellow="\e[0;33m"
+reset="\e[0m"
+normal=$(tput sgr0)
 SRC_DIR=$PWD
 CONFIG=$SRC_DIR/.config
 SOURCE_DIR=$SRC_DIR/source
@@ -25,23 +30,23 @@ ISOIMAGE_OVERLAY=$WORK_DIR/isoimage_overlay
 # shells but it is not POSIX compliant.
 read_property() (
   # The property we are looking for.
-  prop_name=$1
+  prop_name="$1"
 
   # The value of the property set initially to empty string.
   prop_value=
 
   if [ ! "$prop_name" = "" ] ; then
     # Search in the main '.config' file.
-    prop_value=`grep -i ^${prop_name}= $CONFIG | cut -f2- -d'=' | xargs`
+    prop_value="$(grep -i ^${prop_name}= $CONFIG | cut -f2- -d'=' | xargs)"
   fi
 
   echo $prop_value
 )
 
 # Read commonly used properties from the main '.config' file.
-JOB_FACTOR=`read_property JOB_FACTOR`
-CFLAGS=`read_property CFLAGS`
-NUM_CORES=$(grep ^processor /proc/cpuinfo | wc -l)
+JOB_FACTOR="$(read_property JOB_FACTOR)"
+CFLAGS="$(read_property CFLAGS)"
+NUM_CORES="$(grep -c ^processor /proc/cpuinfo)"
 
 # Calculate the number of 'make' jobs to be used later.
 NUM_JOBS=$((NUM_CORES * JOB_FACTOR))
@@ -50,19 +55,19 @@ download_source() (
   url=$1  # Download from this URL.
   file=$2 # Save the resource in this file.
 
-  local=`read_property USE_LOCAL_SOURCE`
+  local="$(read_property USE_LOCAL_SOURCE)"
 
-  if [ "$local" = "true" -a ! -f $file  ] ; then
-    echo "Source file '$file' is missing and will be downloaded."
+  if [ "$local" = "true" -a ! -f "$file"  ] ; then
+    echo "${bold}Source file '$file' is missing and will be downloaded.${normal}"
     local=false
   fi
 
   if [ ! "$local" = "true" ] ; then
-    echo "Downloading source file from '$url'."
-    echo "Saving source file in '$file'".
-    wget -O $file -c $url
+    echo "${bold}Downloading source file from '$url'.${normal}"
+    echo "${bold}Saving source file in '$file'.${normal}"
+    wget -O "$file" -c "$url"
   else
-    echo "Using local source file '$file'."
+    echo "${bold}Using local source file '$file'.${normal}"
   fi
 )
 
@@ -71,10 +76,14 @@ extract_source() (
   name=$2
 
   # Delete folder with previously extracted source.
-  echo "Removing '$name' work area. This may take a while."
-  rm -rf $WORK_DIR/$name
-  mkdir $WORK_DIR/$name
+  echo "${bold}Removing '$name' work area. This may take a while.${normal}"
+  rm -rf "${WORK_DIR:?}"/"$name"
+  mkdir "${WORK_DIR:?}"/"$name"
 
   # Extract source to folder 'work/$source'.
-  tar -xvf $file -C $WORK_DIR/$name
+  tar -xvf "$file" -C "${WORK_DIR:?}"/"$name"
 )
+
+info_print() {
+  echo "${yellow}*${reset} ${blue}$*${reset}"
+}
